navbar working
